requirements.txt
Django>=5.2.8
djangorestframework>=3.14.0
Pillow>=9.0.0
channels>=4.0
channels-redis>=4.0.0
django-cors-headers>=4.0.0